+++ SDK/RMVision/RMVision/src/RMFakeVision.mm
@@ -11,6 +11,7 @@
 #import "RMVisionModule.h"
 #import "RMVisionDebugBroker.h"
 #import <AVFoundation/AVFoundation.h>
+#import "RMVision-Prefix.pch"
 
 @interface RMFakeVision ()
 
@@ -348,7 +349,7 @@
 
 - (BOOL)activateModuleWithName:(NSString *)moduleName
 {
-    int prevCount = self.activeModules.count;
+    NSInteger prevCount = self.activeModules.count;
     BOOL retVal = [super activateModuleWithName:moduleName];
     if (_captureStarted && prevCount == 0 && self.activeModules.count > 0) {
         // we'd better start things up
