+++ SDK/RMVision/RMVision/src/Utils/RMOpenCVUtils.mm
@@ -21,8 +21,8 @@
     size_t outputRow = 0;
     for (size_t i = 0; i < mat.rows; i++)
     {
-        if (mask.at<uchar>(i)) {
-            mat.row(i).copyTo(output.row(outputRow));
+        if (mask.at<uchar>((int)i)) {
+            mat.row((int)i).copyTo(output.row((int)outputRow));
             outputRow++;
         }
     }
@@ -56,17 +56,17 @@
     }
 
     size_t totalPixels = image.rows * image.cols;
-    cv::Mat tmp = image.reshape(1, totalPixels);
+    cv::Mat tmp = image.reshape(1, (int)totalPixels);
     cv::Mat maskVector = localMask.reshape(1, localMask.rows*localMask.cols);
     
     size_t nonZeroCount = cv::countNonZero(maskVector);
     
-    imageVector.create(nonZeroCount, tmp.cols, tmp.type());
+    imageVector.create((int)nonZeroCount, tmp.cols, tmp.type());
     
     size_t imageVectorRow = 0;
     for (size_t i = 0; i < tmp.rows; i++) {
-        if (maskVector.at<uchar>(i)) {
-            tmp.row(i).copyTo(imageVector.row(imageVectorRow));
+        if (maskVector.at<uchar>((int)i)) {
+            tmp.row((int)i).copyTo(imageVector.row((int)imageVectorRow));
             imageVectorRow++;
         }
     }
